Make sure to not create a response wrapepr object if not present
If the exception is a conneciton error (e.g. Timeout, SSL error), the response object ends up being nil and it doesn't make sense to wrap a nil object. In case of network/connection error, just skip wrapping it so we can save work and things will be less buggy.